scraper/requirements.txt
httpx[http2]>=0.27
feedparser>=6.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12
lxml>=5.1
apscheduler>=3.10
psycopg2-binary>=2.9
sqlalchemy>=2.0
openai>=1.30
python-dotenv>=1.0
pydantic>=2.7
